social_media/requirements.txt
Django>=3.2,<4.0
djangorestframework
django-cors-headers
pillow
djangorestframework_simplejwt